requirements.txt
django==2.2.25
djangorestframework==3.12.0
django-rest-auth==0.9.5
django-allauth==0.41.0
coreapi==2.3.3
pyyaml==5.4
django-rest-swagger==2.2.0
django-environ==0.8.1
django-filter==21.1
django-crispy-forms==1.13.0
